We can afford the exra cycles to use double precision floats here
;)

+++ src/time.rs
@@ -37,8 +37,8 @@ impl<'a> Timer<'a> {
     fn print_elapsed(&mut self) {
         if self.output {
             let elapsed = self.elapsed();
-            let time = (elapsed.as_secs() as f32) * 1e3
-                       + (elapsed.subsec_nanos() as f32) / 1e6;
+            let time = (elapsed.as_secs() as f64) * 1e3
+                       + (elapsed.subsec_nanos() as f64) / 1e6;
             let stderr = io::stderr();
             // Arbitrary output format, subject to change.
             writeln!(stderr.lock(),
